requirements.txt
pytest==7.3.1
allure-pytest==2.13.1
selenium==4.9.0
